Created Views, more info in PullRequest description
