+++ TestApp/TestApp/ViewModel/TeacherCreateViewModel.cs
@@ -3,21 +3,22 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TestApp.Model;
 
-namespace TestApp
+namespace TestApp.ViewModel
 {
-    public class TeacherStudentViewModel
+    public class TeacherCreateViewModel
     {
         #region Constant Fields
         #endregion
 
         #region Fields
-        private static TeacherStudentViewModel instance = null;
+        private static TeacherCreateViewModel instance = null;
         private static readonly object padlock = new object();
         #endregion
 
         #region Constructors
-        public TeacherStudentViewModel()
+        public TeacherCreateViewModel()
         {
             //Created but left empty intentionally in case it will be used in the future
         }
@@ -30,7 +31,7 @@ namespace TestApp
         #endregion
 
         #region Properties
-        public static TeacherStudentViewModel Instance
+        public static TeacherCreateViewModel Instance
         {
             get
             {
@@ -38,24 +39,34 @@ namespace TestApp
                 {
                     if (instance==null)
                     {
-                        instance = new TeacherStudentViewModel();
+                        instance = new TeacherCreateViewModel();
                     }
                     return instance;
                 }
             }
         }
+        public Test CreatedTest { get; set; }
+        public Question CreatedQuestion { get; set; }
+        public List<Question> SubjectQuestions { get; set; }
+
         #endregion
 
         #region Methods
-        public void GetStudents()
+        public void CreateTest()
         {
             throw new NotImplementedException();
         }
 
-        public void DisplayStudentResult()
+        public void CreateQuestion()
         {
             throw new NotImplementedException();
         }
+
+        public void GetQuestionsForTestCreation()
+        {
+            throw new NotImplementedException();
+        }
+
         #endregion
 
     }
