+++ TestApp/TestApp/ViewModel/TeacherGradeTestViewModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TestApp
+namespace TestApp.ViewModel
 {
     public class TeacherGradeTestViewModel
     {
